clean: simplify and expose reportOnShake
